tests/wpt/web-platform-tests/bluetooth/requestDevice/acceptAllDevices/optional-services-missing.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = 'requestDevice called with acceptAllDevices: true and ' +
    'with no optionalServices. Should not get access to any services.';
const expected = new DOMException(
    'Origin is not allowed to access any service. ' +
    'Tip: Add the service UUID to \'optionalServices\' in ' +
    'requestDevice() options. https://goo.gl/HxfxSQ',
    'SecurityError');

bluetooth_test(() => getConnectedHealthThermometerDevice({acceptAllDevices: true})
    .then(({device}) => assert_promise_rejects_with_message(
      device.gatt.getPrimaryServices(),
      expected)),
    test_desc);
</script>
